Adding the template command to ino
Templates are an important starting point for beginners and nice to have
for frequently used schemes.
The arduino IDE is already shipping with a lot of builtin examples. The
most logical thing would be to make those available to ino as well.
Copying these by hand is tedious and error prone. Thus the command
	ino template --update
will try to fetch these from the arduino install directory and put them
in the templates folder.
Further primitive administration routines like add, remove or list are
provided.
This gives the opportunity to keep the examples from arduino up-to-date
in case the IDE is updated and shipped with new examples.

AUTHOR: Tobias Pleyer